refactor(ui): rename components/ to appComponents/
Disambiguate app-level components from the shared @opentrace/components
library by renaming ui/src/components/ to ui/src/appComponents/.
