+++ ui/src/App.tsx
@@ -17,10 +17,10 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { useJobService, useJobStream } from './job';
 import type { JobMessage } from './job';
-import GraphViewer from './components/GraphViewer';
-import type { GraphViewerHandle } from './components/GraphViewer';
-import ChatPanel from './components/ChatPanel';
-import SettingsDrawer from './components/SettingsDrawer';
+import GraphViewer from './appComponents/GraphViewer';
+import type { GraphViewerHandle } from './appComponents/GraphViewer';
+import ChatPanel from './appComponents/ChatPanel';
+import SettingsDrawer from './appComponents/SettingsDrawer';
 import type { GraphNode, GraphLink } from '@opentrace/components/utils';
 import { loadAnimationSettings } from './config/animation';
 import { normalizeRepoUrl, detectProvider } from '@opentrace/components';
