+++ ui/src/appComponents/ChatPanel.tsx
