+++ ui/src/appComponents/DiscoverPanelContainer.tsx
