+++ ui/src/appComponents/EdgeDetailsPanel.css
