+++ ui/src/appComponents/EdgeDetailsPanel.tsx
