+++ ui/src/appComponents/JobMinimizedBar.css
