+++ ui/src/appComponents/JobMinimizedBar.tsx
