+++ ui/src/appComponents/NodeDetailsPanel.css
