+++ ui/src/appComponents/NodeDetailsPanel.tsx
