+++ ui/src/appComponents/OpenTraceLogo.tsx
