+++ ui/src/appComponents/PRDetailPanel.css
