+++ ui/src/appComponents/PRDetailPanel.tsx
