+++ ui/src/appComponents/ResetConfirmModal.tsx
