+++ ui/src/appComponents/SidePanel.css
