+++ ui/src/appComponents/SidePanel.tsx
