+++ ui/src/appComponents/providerIcons.tsx
